trunk/tests/test_Freja-History.html

<html>
<head>
    <!-- MochiKit is needed by SimpleTest -->
    <script type="text/javascript" src="../lib/mochi+sarissa/MochiKit.js"></script>
    <script type="text/javascript" src="SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="../lib/mochi+sarissa/Sarissa.js"></script>
    <script type="text/javascript" src="../lib/mochi+sarissa/Freja.js"></script>
    <link rel="stylesheet" type="text/css" href="SimpleTest/test.css">
</head>
<body>

<pre id="test">
<script type="text/javascript" src="test_History.js"></script>
<script type="text/javascript">
try {

    tests.test_History({ok:ok, is:is});
    ok( true, "test suite finished!");

} catch (err) {

    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok ( false, s );

}
</script>
</pre>
</body>
</html>
